backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
lancedb==0.30.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pymupdf==1.25.5
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic[email]==2.9.2
